% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidybulk_functions.R
\name{ciber_deconvolve}
\alias{ciber_deconvolve}
\title{Perform cibersort deconvolution on tidybulk object}
\usage{
ciber_deconvolve(tb_obj)
}
\arguments{
\item{tb_object}{tidybulk object
returns dataframe with cibersort deconvolution results}
}
\description{
Perform cibersort deconvolution on tidybulk object
}
