% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidybulk_functions.R
\name{get_tidybulk_topgenes_pval}
\alias{get_tidybulk_topgenes_pval}
\title{return n top differentially expressed gene symbols from tidybulk differential gene expression results}
\usage{
get_tidybulk_topgenes_pval(counts_de, n_top = 10)
}
\arguments{
\item{counts_de}{tidybulk differential expression result}

\item{n_top}{return the n_top gene symbols}
}
\description{
return n top differentially expressed gene symbols from tidybulk differential gene expression results
}
