% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidybulk_functions.R
\name{make_abundance_density_plot}
\alias{make_abundance_density_plot}
\title{returns a ggplot2 showing density plots of raw and scaled abundance counts from tidybulk object}
\usage{
make_abundance_density_plot(
  tb_obj,
  raw_count_col,
  scaled_count_col,
  sampleID_col
)
}
\arguments{
\item{tb_obj}{tidybulk object with raw and scaled counts}

\item{raw_count_col}{column name contaning raw counts}

\item{scaled_count_col}{column name containing scaled counts}

\item{sampleID_col}{column name name listing containing sampleIDs of analysis}
}
\description{
returns a ggplot2 showing density plots of raw and scaled abundance counts from tidybulk object
}
