% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidybulk_functions.R
\name{make_ciber_paired_box_point_plot}
\alias{make_ciber_paired_box_point_plot}
\title{return ggplot2 paired plot of cibersort cell fractions Pre_treatment Post_treatment for paired samples}
\usage{
make_ciber_paired_box_point_plot(ciber_res_paired_df, cell_type_list)
}
\arguments{
\item{ciber_res_paired_df}{dataframe retured from}

\item{cell_type_list}{list of cell types to analyze}
}
\description{
return ggplot2 paired plot of cibersort cell fractions Pre_treatment Post_treatment for paired samples
}
