% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidybulk_functions.R
\name{make_ciber_plot}
\alias{make_ciber_plot}
\title{makes stacked bar plot showing cell type proportions of sampleIDs, faceted by subjectID}
\usage{
make_ciber_plot(ciber_res, category_factor, fill_factor)
}
\arguments{
\item{ciber_res}{is datframe of cibersort results returned from ciber_deconvolve}

\item{category_factor}{for plotting cell proportions  (e.g Treated)}

\item{fill}{factor to fill the stacked proportions
returns ggplot2 object}
}
\description{
makes stacked bar plot showing cell type proportions of sampleIDs, faceted by subjectID
}
