% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidybulk_functions.R
\name{make_pca_plot}
\alias{make_pca_plot}
\title{returns PCA plot of tidybulk abundance scaled object}
\usage{
make_pca_plot(tb_obj, color_factor, shape_factor, label_col, dim1, dim2)
}
\arguments{
\item{tb_obj}{tidybulk object with raw and scaled counts}

\item{color_factor}{for PCA plot data points}

\item{shape_factor}{for PC plot data points}

\item{label_col}{for labelling points on data points}

\item{dim1}{PC for x-axis}

\item{dim2}{PC for y-axis}
}
\description{
returns PCA plot of tidybulk abundance scaled object
}
