% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidybulk_functions.R
\name{make_tcon_treg_proportions_plot}
\alias{make_tcon_treg_proportions_plot}
\title{make scatter plot of Tcon vs Treg cell type proproportions from cibersort results}
\usage{
make_tcon_treg_proportions_plot(
  tcon_treg_table,
  shape_factor,
  color_factor,
  subtitle_string = ""
)
}
\arguments{
\item{tcon_treg_table}{from make_tcon_treg_proportions_table}

\item{shape_factor}{name of column that you want to map shape of points to in plot}

\item{color_factor}{name of column you want to map color of points to in plot}
}
\description{
make scatter plot of Tcon vs Treg cell type proproportions from cibersort results
}
