% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidybulk_functions.R
\name{make_tcon_treg_proportions_table}
\alias{make_tcon_treg_proportions_table}
\title{return dataframe Tcon/Treg cell type proportions}
\usage{
make_tcon_treg_proportions_table(
  ciber_res,
  category_factor,
  subject_column,
  sample_column
)
}
\arguments{
\item{ciber_res}{dataframe with tidybulk cibersort results}

\item{category_factor}{column representing with category levels you want Treg/Tcon cell proportions for}

\item{subject_column}{column representing with subjectID}

\item{sample_column}{column representing with sample name}
}
\description{
return dataframe Tcon/Treg cell type proportions
}
