% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidybulk_functions.R
\name{make_tidybulk}
\alias{make_tidybulk}
\title{returns a tidybulk object from dataframe (tibble)}
\usage{
make_tidybulk(df, sample_col, transcript_col, abundance_col)
}
\arguments{
\item{df}{Dataframe/tibble with at least 3 columns for sampleID, gene symbol, and transcript abundance}

\item{sample_col}{The name of the sample column}

\item{transcript_col}{The name of the gene/transcript column}

\item{abundance_col}{The name of the gene/transcript abundance column (raw counts)}
}
\description{
returns a tidybulk object from dataframe (tibble)
}
