% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidybulk_functions.R
\name{make_tidybulk_volcano_plot}
\alias{make_tidybulk_volcano_plot}
\title{return ggplot volcano plot from tidybulk differential expression result}
\usage{
make_tidybulk_volcano_plot(
  counts_de,
  pval_limit = 0.05,
  FC_lim = 1,
  title_string = "Tidybulk volcano plot",
  subtitle_string = ""
)
}
\arguments{
\item{counts_de}{tidybulk differential expression result}

\item{FC_lim}{fold change default is 2}

\item{top_gene_symbols}{list of gene symbols to label in volcano plot}

\item{FDR_lim}{default is 0.05}
}
\description{
return ggplot volcano plot from tidybulk differential expression result
}
