% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidybulk_functions.R
\name{make_tidybulk_volcano_plot_top_genes}
\alias{make_tidybulk_volcano_plot_top_genes}
\title{return ggplot volcano plot from tidybulk differential expression result, labeling gene_symbols}
\usage{
make_tidybulk_volcano_plot_top_genes(
  counts_de,
  gene_symbols,
  FC_lim = 1,
  title_string = "Tidybulk volcano plot",
  subtitle_string = ""
)
}
\arguments{
\item{counts_de}{tidybulk differential expression result}

\item{gene_symbols}{list of gene symbols to label in volcano plot}

\item{FC_lim}{fold change default is 2}

\item{FDR_lim}{default is 0.05}
}
\description{
return ggplot volcano plot from tidybulk differential expression result, labeling gene_symbols
}
